\name{print.scalreg}
\alias{print.scalreg}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Printing the solution from a scalreg object
}
\description{
Print the solution from a scalreg object
}
\usage{
\method{print}{scalreg}(x,...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{a scalreg object}
  \item{\dots}{Additonal arguments for generic methods}  	
}


\author{
Tingni Sun}

\seealso{
scalreg
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{print}
%\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
