\name{scalreg-internal}
\title{Internal scalreg functions}
\alias{slassoEst}
\alias{slassoInv}
\alias{lse}

\description{Internal scalreg functions}
\usage{
slassoEst(X,y,lam0=NULL)
slassoInv(X,lam0=NULL,LSE=F)
lse(X,y,indexset)
}
\author{
Tingni Sun
}
\details{
  These are not to be called by the user.
}
\keyword{internal}
