\name{scalreg-package}
\alias{scalreg-package}
\docType{package}
\title{
Scaled sparse linear regression
}
\description{
This package fits scaled sparse linear regression with l_1 penalty. The algorithm jointly estimates the regression coefficients and the noise level in linear regression problem. In addition, the package estimates inverse covariance matrices (precision matrices) via a scale-invariant method.
}
\details{
\tabular{ll}{
Package: \tab scalreg\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-12-16\cr
License: \tab GPL-2\cr
}
%%~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Tingni Sun <tingni@wharton.upenn.edu>
}
\references{
Sun, T. and Zhang, C.-H. (2012) Scaled sparse linear regression. Biometrika, 99 (4), 879-898.

Sun, T. and Zhang, C.-H. (2013) Sparse matrix inversion with scaled Lasso. Journal of Machine Learning Research, 14, 3385-3418.
}
\keyword{ package }
\seealso{
scalreg
}
\examples{
## See examples in scalreg}
